backend/requirements.txt
absl-py==2.1.0
affine==2.4.0
altair==5.4.1
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==24.2.0
backcall==0.2.0
basemap==1.4.1
basemap-data==1.3.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
bokeh==3.6.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contextily==1.6.2
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.35
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastjsonschema==2.16.3
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.10.0
gast==0.6.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.35.1
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.18.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lxml==4.9.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.18
mercantile==1.2.1
mistune==2.0.5
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.9.4
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.3
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.4
openai==1.63.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
outcome==1.2.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.5.0
plotly==5.24.1
pmdarima==2.0.3
portpicker==1.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.5
psutil==5.9.5
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydomo==0.3.0.9
Pygments==2.18.0
pyogrio==0.10.0
pyopencl==2024.3
pyparsing==3.1.2
pyproj==3.6.1
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-json-logger==2.0.7
pytools==2024.1.14
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
rasterio==1.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.9.0
Send2Trash==1.8.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.4.1
srt==3.5.3
stack-data==0.6.2
starlette==0.45.3
statsmodels==0.14.0
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2024.9.20
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.3.1
tqdm==4.66.5
traitlets==5.9.0
transformers==4.48.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.20
uvicorn==0.34.0
vega-datasets==0.9.0
vl-convert-python==1.7.0
vosk==0.3.45
watchdog==5.0.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==14.2
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.9.0
zipp==3.21.0
